more folder rename shenanigans
